# openapi.yaml
openapi: "3.1.0"
info:
  title: Gregor Zwanzig API
  description: Go REST API for weather risk reports for long-distance hiking.
  version: "0.1.0"
  contact:
    name: Henning Emmrich
servers:
  - url: https://gregor20.henemm.com
    description: Production
  - url: http://localhost:8090
    description: Local development
paths:
...
  /api/trips/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          example: gr221-mallorca
    get:
      summary: Get a trip by ID
      operationId: getTrip
      tags: [trips]
      responses:
        "200":
          description: Trip found
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Update a trip
      operationId: updateTrip
      tags: [trips]
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Trip"
      responses:
        "200":
          description: Trip updated
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "412":
          description: >-
            If-Match stimmt nicht mit dem aktuellen Stand ueberein — der Trip
            wurde zwischenzeitlich an anderer Stelle geaendert. Es wurde nichts
            geschrieben.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreconditionFailedError"
    delete:
      summary: Delete a trip
      operationId: deleteTrip
      tags: [trips]
      responses:
        "204":
          description: Trip deleted (idempotent)
...
components:
  headers:
    ETag:
      description: >-
        Inhalts-Fingerabdruck (sha256-Hex) der gespeicherten Briefing-Datei,
        in doppelten Anfuehrungszeichen (RFC 7232). Als If-Match an einen
        Schreibvorgang zurueckgeben, um verlorene Aenderungen zu verhindern
        (Issue #1395). Fehlt der Header, ist der Fingerabdruck gerade nicht
        lesbar — der Aufruf bleibt gueltig, nur ohne Vorbedingungsschutz.
      schema:
        type: string
        example: '"a1b2c3d4e5f60718293a4b5c6d7e8f90a1b2c3d4e5f60718293a4b5c6d7e8f90"'
  parameters:
    IfMatch:
      name: If-Match
      in: header
      required: false
      description: >-
        Zuletzt erhaltener ETag-Wert. Optional: ohne diesen Header wird der
        Schreibvorgang unveraendert angenommen (Rollout-Politik Issue #1395 S2).
        Stimmt der Wert nicht mit dem aktuellen Stand ueberein, antwortet der
        Server mit 412 und schreibt nichts. Erlaubt sind eine kommagetrennte
        Liste (ein Treffer genuegt) und "*".
      schema:
        type: string
        example: '"a1b2c3d4e5f60718293a4b5c6d7e8f90a1b2c3d4e5f60718293a4b5c6d7e8f90"'
  schemas:
...
    Waypoint:
      type: object
      required: [id, name, lat, lon, elevation_m]
      properties:
        id:
          type: string
          example: W1
        name:
          type: string
          example: Start
        lat:
          type: number
          format: double
          example: 47.0
        lon:
          type: number
          format: double
          example: 11.0
        elevation_m:
          type: integer
          example: 1200
        time_window:
          type: string
          description: "Format: HH:MM-HH:MM"
          example: "08:00-10:00"
    Stage:
      type: object
      required: [id, name, date, waypoints]
      properties:
        id:
          type: string
          example: S1
        name:
          type: string
          example: "Tag 1"
        date:
          type: string
          format: date
          example: "2026-04-13"
        waypoints:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/Waypoint"
        start_time:
          type: string
          description: "Format: HH:MM:SS"
          example: "09:00:00"
    Trip:
      type: object
      required: [id, name, stages]
      properties:
        id:
          type: string
          example: gr221-mallorca
        name:
          type: string
          example: GR221 Mallorca
        stages:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/Stage"
        avalanche_regions:
          type: array
          items:
            type: string
        aggregation:
          type: object
          description: Aggregation configuration
        weather_config:
          type: object
          description: Legacy weather config (opaque)
        display_config:
          type: object
          description: Weather display configuration (opaque)
        report_config:
          type: object
          description: Report scheduling configuration (opaque)
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: not_found
    ValidationError:
      type: object
      required: [error, detail]
      properties:
        error:
          type: string
          example: validation_error
        detail:
          type: string
          example: name required
    PreconditionFailedError:
      type: object
      required: [error, detail]
      properties:
        error:
          type: string
          enum: [precondition_failed]
          example: precondition_failed
        detail:
          type: string
          description: Nutzerlesbare deutsche Erklaerung inkl. Handlungsempfehlung
          example: >-
            Der Stand wurde zwischenzeitlich an anderer Stelle geaendert. Bitte
            neu laden und die Aenderung erneut vornehmen.